Fix merged CI type and lint failures
